+++ src/cqterrain/bridge/__init__.py
@@ -3,3 +3,4 @@ from .BaseRamp import BaseRamp
 from .Bridge import Bridge
 
 from .TileStraight import TileStraight
+from .TileBridge import TileBridge